refactor: name modules with intake names
